refactor(analysis): decompose loader.py into focused sub-modules
Split src/scylla/analysis/loader.py (933 LOC) into four focused modules
under src/scylla/analysis/loader_internals/:

  - validators.py         (118 LOC) — JSON scalar coercion helpers
  - models.py             (194 LOC) — dataclasses + RubricConflict types
  - run_loader.py         (418 LOC) — per-run JSON parsing + judge model
  - experiment_loader.py  (229 LOC) — directory traversal + rubric merging

loader.py is now a 92-line facade re-exporting every previously public name
for backward compatibility (no caller needs to change). Move-only refactor;
no behavior changes.

Partial fix for #1876 — three other oversized files (e2e/models.py,
e2e/runner.py, e2e/tier_manager.py) still need their own decomposition PRs.

Refs #1876
