+++ src/scylla/analysis/loader.py
@@ -1,933 +1,92 @@
 """Data loader for experiment results.
 
 Loads experiment data from the fullruns/ directory hierarchy and converts
-to structured dataclasses for analysis.
-data structures. Uses existing e2e.models data structures to avoid duplication.
+to structured dataclasses for analysis. Uses existing e2e.models data
+structures to avoid duplication.
+
+This module is a thin facade — its implementation lives in
+:mod:`scylla.analysis.loader_internals`. The public API (every name listed
+in ``__all__``) is preserved for backward compatibility.
 """
 
 from __future__ import annotations
 
-import json
-import logging
-import re
-import warnings
-from dataclasses import dataclass
-from pathlib import Path
-from typing import Any, Literal, cast
-
-import jsonschema
-import numpy as np
-import yaml
-
-from scylla.e2e.models import TokenStats
-
-logger = logging.getLogger(__name__)
-
-# Rubric conflict resolution policy.
-# 'error'  – raise RubricConflictError (default; safest for research pipelines)
-# 'warn'   – emit UserWarning and keep the *first* value
-# 'first'  – silently keep the first value encountered
-# 'last'   – silently overwrite with the last value encountered
-RubricConflict = Literal["error", "warn", "first", "last"]
-
-
-class RubricConflictError(ValueError):
-    """Raised when two experiments define conflicting rubric weights for the same category.
-
-    Attributes:
-        category: Category name where the conflict was detected.
-        exp_first: Experiment name that first defined this category.
-        weight_first: Weight from the first experiment.
-        exp_second: Experiment name that introduced the conflict.
-        weight_second: Conflicting weight from the second experiment.
-
-    """
-
-    def __init__(
-        self,
-        category: str,
-        exp_first: str,
-        weight_first: float,
-        exp_second: str,
-        weight_second: float,
-    ) -> None:
-        """Initialize RubricConflictError with conflict details."""
-        self.category = category
-        self.exp_first = exp_first
-        self.weight_first = weight_first
-        self.exp_second = exp_second
-        self.weight_second = weight_second
-        super().__init__(
-            f"Rubric conflict for category '{category}': "
-            f"experiment '{exp_first}' defines weight={weight_first}, "
-            f"but experiment '{exp_second}' defines weight={weight_second}. "
-            "Use the rubric_conflict parameter to control resolution."
-        )
-
-
-# Load JSON Schema for run_result.json validation
-_SCHEMA_PATH = Path(__file__).parent / "schemas" / "run_result.schema.json"
-with _SCHEMA_PATH.open() as _schema_file:
-    _RUN_RESULT_SCHEMA = json.load(_schema_file)
-
-
-def validate_numeric(value: Any, field_name: str, default: float = np.nan) -> float:
-    """Validate and coerce numeric field from JSON.
-
-    Args:
-        value: Value from JSON (could be int, float, str, None, etc.)
-        field_name: Field name for error messages
-        default: Default value if validation fails
-
-    Returns:
-        Validated float value or default
-
-    """
-    # Handle None or missing
-    if value is None:
-        return default
-
-    # Try to convert to float
-    try:
-        result = float(value)
-
-        # Check for invalid values
-        if np.isnan(result) or np.isinf(result):
-            return default
-
-        return result
-    except (ValueError, TypeError):
-        logger.warning(
-            "Invalid type for %s: %s (value=%r), using default=%s",
-            field_name,
-            type(value).__name__,
-            value,
-            default,
-        )
-        return default
-
-
-def validate_bool(value: Any, field_name: str, default: bool = False) -> bool:
-    """Validate and coerce boolean field from JSON.
-
-    Args:
-        value: Value from JSON (could be bool, int, str, None, etc.)
-        field_name: Field name for error messages
-        default: Default value if validation fails
-
-    Returns:
-        Validated bool value or default
-
-    """
-    # Handle None or missing
-    if value is None:
-        return default
-
-    # If already bool, return directly
-    if isinstance(value, bool):
-        return value
-
-    # Try common string representations
-    if isinstance(value, str):
-        value_lower = value.lower()
-        if value_lower in ("true", "yes", "1"):
-            return True
-        if value_lower in ("false", "no", "0"):
-            return False
-
-    # Try numeric conversion (0 = False, non-zero = True)
-    try:
-        return bool(int(value))
-    except (ValueError, TypeError):
-        logger.warning(
-            "Invalid type for %s: %s (value=%r), using default=%s",
-            field_name,
-            type(value).__name__,
-            value,
-            default,
-        )
-        return default
-
-
-def validate_int(value: Any, field_name: str, default: int = -1) -> int:
-    """Validate and coerce integer field from JSON.
-
-    Args:
-        value: Value from JSON (could be int, float, str, None, etc.)
-        field_name: Field name for error messages
-        default: Default value if validation fails
-
-    Returns:
-        Validated int value or default
-
-    """
-    # Handle None or missing
-    if value is None:
-        return default
-
-    # Try to convert to int
-    try:
-        return int(value)
-    except (ValueError, TypeError):
-        logger.warning(
-            "Invalid type for %s: %s (value=%r), using default=%s",
-            field_name,
-            type(value).__name__,
-            value,
-            default,
-        )
-        return default
-
-
-@dataclass
-class CriterionScore:
-    """Detailed score for a single rubric criterion.
-
-    Attributes:
-        name: Criterion name (functional, code_quality, etc.)
-        achieved: Points achieved
-        max_points: Maximum possible points
-        score: Normalized score (0.0-1.0)
-        items: Individual check items within this criterion
-
-    """
-
-    name: str
-    achieved: float
-    max_points: float
-    score: float
-    items: dict[str, ItemScore]
-
-
-@dataclass
-class ItemScore:
-    """Score for an individual rubric check item.
-
-    Attributes:
-        item_id: Item identifier (e.g., "F1", "Q2")
-        achieved: Points achieved (or "N/A")
-        max_points: Maximum possible points (or "N/A")
-        reason: Judge's reasoning for the score
-
-    """
-
-    item_id: str
-    achieved: float | str
-    max_points: float | str
-    reason: str
-
-
-@dataclass
-class JudgeEvaluation:
-    """A single judge's evaluation of a run.
-
-    Attributes:
-        judge_model: Model ID of the judge (e.g., "claude-opus-4-6")
-        judge_number: Judge number (1, 2, or 3)
-        score: Overall score (0.0-1.0)
-        passed: Whether the run passed
-        grade: Letter grade (S/A/B/C/D/F)
-        is_valid: Whether the judgment was valid
-        reasoning: Judge's overall reasoning
-        criteria: Detailed scores by criterion
-
-    """
-
-    judge_model: str
-    judge_number: int
-    score: float
-    passed: bool
-    grade: str
-    is_valid: bool
-    reasoning: str
-    criteria: dict[str, CriterionScore]
-
-
-@dataclass
-class ModelUsage:
-    """Per-model token usage from agent execution.
-
-    Tracks individual model usage when multiple models are involved
-    (relevant for T3-T5 delegation tiers).
-
-    Attributes:
-        model: Model identifier
-        input_tokens: Input tokens consumed
-        output_tokens: Output tokens generated
-        cache_creation_tokens: Cache creation tokens
-        cache_read_tokens: Cache read tokens
-        cost_usd: Cost for this model's usage
-
-    """
-
-    model: str
-    input_tokens: int
-    output_tokens: int
-    cache_creation_tokens: int = 0
-    cache_read_tokens: int = 0
-    cost_usd: float = 0.0
-
-
-@dataclass
-class RunData:
-    """Complete data for a single run.
-
-    Attributes:
-        experiment: Experiment identifier
-        agent_model: Model used for agent (claude-sonnet-4-6, claude-haiku-4-5)
-        tier: Tier ID (T0-T6)
-        subtest: Subtest ID
-        run_number: Run number (1-10)
-        score: Consensus judge score (arithmetic mean of judges)
-        passed: Consensus pass decision (majority vote)
-        grade: Consensus grade
-        cost_usd: Total cost in USD
-        duration_seconds: Total duration
-        agent_duration_seconds: Agent execution time
-        judge_duration_seconds: Judge evaluation time
-        token_stats: Detailed token usage
-        exit_code: Agent exit code
-        judges: Per-judge evaluations
-        api_calls: Number of API calls (optional, for delegation tiers)
-        num_turns: Number of agentic turns (optional, for delegation tiers)
-        model_usage: Per-model token usage (optional, for delegation tiers)
-        r_prog: Fine-Grained Progress Rate, 0.0-1.0 (optional, from process_metrics)
-        strategic_drift: Strategic Drift score, 0.0-1.0 (optional, from process_metrics)
-        cfp: Change Fail Percentage, 0.0-1.0 (optional, from process_metrics)
-        pr_revert_rate: PR Revert Rate, 0.0-1.0 (optional, from process_metrics)
-
-    """
-
-    experiment: str
-    agent_model: str
-    tier: str
-    subtest: str
-    run_number: int
-    score: float
-    passed: bool
-    grade: str
-    cost_usd: float
-    duration_seconds: float
-    agent_duration_seconds: float
-    judge_duration_seconds: float
-    token_stats: TokenStats
-    exit_code: int
-    judges: list[JudgeEvaluation]
-    # Optional agent result fields (from agent/result.json)
-    api_calls: int | None = None
-    num_turns: int | None = None
-    model_usage: list[ModelUsage] | None = None
-    # Optional process metrics (from run_result.json process_metrics block)
-    r_prog: float | None = None
-    strategic_drift: float | None = None
-    cfp: float | None = None
-    pr_revert_rate: float | None = None
-
-
-def model_id_to_display(model_id: str) -> str:
-    """Return model ID as its display name.
-
-    Args:
-        model_id: Model ID (e.g., "claude-sonnet-4-6")
-
-    Returns:
-        The model ID unchanged — display matches the identifier for clarity.
-
-    Examples:
-        >>> model_id_to_display("claude-sonnet-4-6")
-        'claude-sonnet-4-6'
-        >>> model_id_to_display("claude-haiku-4-5")
-        'claude-haiku-4-5'
-        >>> model_id_to_display("unknown-model")
-        'unknown-model'
-
-    """
-    return model_id
-
-
-def _find_model_md_in_experiment(experiment_dir: Path) -> str | None:
-    """Scan tier/subtest/run directories for the first valid agent MODEL.md.
-
-    Args:
-        experiment_dir: Path to experiment directory.
-
-    Returns:
-        Model display string, or None if no MODEL.md found.
-
-    """
-    from scylla.e2e.paths import COMPLETED_DIR
-
-    completed_dir = experiment_dir / COMPLETED_DIR
-    scan_base = completed_dir if completed_dir.exists() else experiment_dir
-    for tier_dir in sorted(scan_base.iterdir()):
-        if not tier_dir.is_dir() or not tier_dir.name.startswith("T"):
-            continue
-        for subtest_dir in sorted(tier_dir.iterdir()):
-            if not subtest_dir.is_dir() or not subtest_dir.name.isdigit():
-                continue
-            for run_dir in sorted(subtest_dir.iterdir()):
-                if not run_dir.is_dir() or not run_dir.name.startswith("run_"):
-                    continue
-                model_md = run_dir / "agent" / "MODEL.md"
-                if model_md.exists():
-                    try:
-                        return model_id_to_display(parse_judge_model(model_md))
-                    except Exception as e:
-                        logger.warning("Failed to parse %s: %s", model_md, e)
-    return None
-
-
-def resolve_agent_model(experiment_dir: Path) -> str:
-    """Resolve agent model from experiment configuration.
-
-    Tries (in order):
-    1. config/experiment.json -> models[0]
-    2. First agent/MODEL.md file found
-
-    Args:
-        experiment_dir: Path to experiment directory (timestamped)
-
-    Returns:
-        Model ID string
-
-    Raises:
-        ValueError: If model cannot be determined
-
-    """
-    # Try experiment.json first
-    config_path = experiment_dir / "config" / "experiment.json"
-    if config_path.exists():
-        try:
-            with config_path.open() as f:
-                config = json.load(f)
-                models = config.get("models", [])
-                if models:
-                    return model_id_to_display(models[0])
-        except Exception as e:
-            logger.warning("Failed to read %s: %s", config_path, e)
-
-    # Fallback: find first agent/MODEL.md
-    model = _find_model_md_in_experiment(experiment_dir)
-    if model is not None:
-        return model
-
-    raise ValueError(f"Could not determine agent model for {experiment_dir}")
-
-
-def parse_judge_model(model_md_path: Path) -> str:
-    """Parse judge model from MODEL.md file.
-
-    Args:
-        model_md_path: Path to MODEL.md file
-
-    Returns:
-        Judge model ID
-
-    Raises:
-        ValueError: If model pattern not found
-
-    """
-    content = model_md_path.read_text()
-    match = re.search(r"\*\*Model\*\*:\s*(.+)", content)
-    if not match:
-        raise ValueError(f"Could not find model in {model_md_path}")
-    return match.group(1).strip()
-
-
-def load_agent_result(run_dir: Path) -> dict[str, Any]:
-    """Load agent execution result from agent/result.json.
-
-    Args:
-        run_dir: Path to the run directory
-
-    Returns:
-        Dictionary with agent result data, or empty dict if not available
-
-    """
-    agent_result_path = run_dir / "agent" / "result.json"
-    if not agent_result_path.exists():
-        return {}
-
-    try:
-        with agent_result_path.open() as f:
-            return cast(dict[str, Any], json.load(f))
-    except (json.JSONDecodeError, OSError) as e:
-        logger.warning("Failed to load agent result %s: %s", agent_result_path, e)
-        return {}
-
-
-def load_judgment(judgment_path: Path, judge_number: int) -> JudgeEvaluation:
-    """Load a single judge's evaluation.
-
-    Args:
-        judgment_path: Path to judgment.json file
-        judge_number: Judge number (1, 2, or 3)
-
-    Returns:
-        Judge evaluation data
-
-    """
-    with judgment_path.open() as f:
-        data = json.load(f)
-
-    # Parse judge model from MODEL.md in same directory
-    model_md_path = judgment_path.parent / "MODEL.md"
-    judge_model = parse_judge_model(model_md_path)
-
-    # Parse criteria scores (handle None case)
-    criteria_scores_data = data.get("criteria_scores")
-    if criteria_scores_data is None:
-        criteria_scores_data = {}
-
-    criteria = {}
-    for criterion_name, criterion_data in criteria_scores_data.items():
-        if criterion_data is None:
-            continue
-
-        items_data = criterion_data.get("items", {})
-        if items_data is None:
-            items_data = {}
-
-        items = {}
-        for item_id, item_data in items_data.items():
-            if item_data is None:
-                continue
-
-            items[item_id] = ItemScore(
-                item_id=item_id,
-                achieved=item_data.get("achieved", "N/A"),
-                max_points=item_data.get("max", "N/A"),
-                reason=item_data.get("reason", ""),
-            )
-
-        criteria[criterion_name] = CriterionScore(
-            name=criterion_name,
-            achieved=criterion_data.get("achieved", np.nan),
-            max_points=criterion_data.get("max", np.nan),
-            score=criterion_data.get("score", np.nan),
-            items=items,
-        )
-
-    # Check is_valid flag
-    is_valid_raw = data.get("is_valid", True) is not False
-
-    return JudgeEvaluation(
-        judge_model=judge_model,
-        judge_number=judge_number,
-        score=data.get("score", np.nan),
-        passed=data.get("passed", False),
-        grade=data.get("grade", "F"),
-        is_valid=is_valid_raw,
-        reasoning=data.get("reasoning", ""),
-        criteria=criteria,
-    )
-
-
-def load_run(run_dir: Path, experiment: str, tier: str, subtest: str, agent_model: str) -> RunData:  # noqa: C901  # data loading with many optional field paths
-    """Load data for a single run.
-
-    Args:
-        run_dir: Path to run directory
-        experiment: Experiment name
-        tier: Tier ID (T0-T6)
-        subtest: Subtest ID
-        agent_model: Agent model display name
-
-    Returns:
-        Complete run data
-
-    """
-    # Load run_result.json for consensus data
-    run_result_path = run_dir / "run_result.json"
-    with run_result_path.open() as f:
-        result = json.load(f)
-
-    # Validate against JSON Schema (graceful degradation - log warning only)
-    try:
-        jsonschema.validate(result, _RUN_RESULT_SCHEMA)
-    except jsonschema.ValidationError as e:
-        logger.warning(
-            "Schema validation failed for %s: %s (path: %s)",
-            run_result_path,
-            e.message,
-            " -> ".join(str(p) for p in e.path) if e.path else "root",
-        )
-
-    # Parse run number from directory name (e.g., "run_01" -> 1)
-    try:
-        run_number = int(run_dir.name.split("_")[1])
-    except (IndexError, ValueError):
-        # Fallback: try to extract any number from the directory name
-        import re
-
-        match = re.search(r"\d+", run_dir.name)
-        run_number = int(match.group()) if match else 0
-
-    # Load token stats
-    token_stats = TokenStats.from_dict(result.get("token_stats", {}))
-
-    # Load per-judge evaluations
-    judges = []
-    judge_dir = run_dir / "judge"
-    if judge_dir.exists():
-        # Dynamically discover all judge directories
-        for judge_path in sorted(judge_dir.glob("judge_*/judgment.json")):
-            # Extract judge number from directory name (e.g., "judge_01" -> 1)
-            judge_num = int(judge_path.parent.name.replace("judge_", ""))
-            judges.append(load_judgment(judge_path, judge_num))
-
-    # Load optional agent result data
-    agent_data = load_agent_result(run_dir)
-    api_calls_val = None
-    num_turns_val = None
-    model_usage_val = None
-
-    if agent_data:
-        # Extract API calls and turns if present
-        if "api_calls" in agent_data:
-            api_calls_val = validate_int(agent_data["api_calls"], "api_calls", 0) or None
-        if "num_turns" in agent_data:
-            num_turns_val = validate_int(agent_data["num_turns"], "num_turns", 0) or None
-
-        # Parse model_usage if present (for delegation tiers)
-        raw_usage = agent_data.get("model_usage") or agent_data.get("modelUsage")
-        if raw_usage and isinstance(raw_usage, list):
-            model_usage_val = []
-            for usage in raw_usage:
-                if isinstance(usage, dict):
-                    model_usage_val.append(
-                        ModelUsage(
-                            model=usage.get("model", "unknown"),
-                            input_tokens=validate_int(
-                                usage.get("input_tokens") or usage.get("inputTokens"),
-                                "input_tokens",
-                                0,
-                            ),
-                            output_tokens=validate_int(
-                                usage.get("output_tokens") or usage.get("outputTokens"),
-                                "output_tokens",
-                                0,
-                            ),
-                            cache_creation_tokens=validate_int(
-                                usage.get("cache_creation_tokens"), "cache_creation_tokens", 0
-                            ),
-                            cache_read_tokens=validate_int(
-                                usage.get("cache_read_tokens"), "cache_read_tokens", 0
-                            ),
-                            cost_usd=validate_numeric(usage.get("cost_usd"), "cost_usd", 0.0),
-                        )
-                    )
-
-    # Extract process metrics from run_result.json
-    r_prog_val: float | None = None
-    strategic_drift_val: float | None = None
-    cfp_val: float | None = None
-    pr_revert_rate_val: float | None = None
-
-    process_metrics_data = result.get("process_metrics")
-    if process_metrics_data and isinstance(process_metrics_data, dict):
-        # Use pre-computed process_metrics block if available
-        def _extract_process_float(key: str) -> float | None:
-            raw = validate_numeric(process_metrics_data.get(key), key, np.nan)
-            return None if (raw is None or np.isnan(raw)) else raw
-
-        r_prog_val = _extract_process_float("r_prog")
-        strategic_drift_val = _extract_process_float("strategic_drift")
-        cfp_val = _extract_process_float("cfp")
-        pr_revert_rate_val = _extract_process_float("pr_revert_rate")
-    else:
-        # Fallback: compute from raw tracking data if present
-        progress_tracking = result.get("progress_tracking")
-        changes = result.get("changes")
-        if progress_tracking or changes:
-            from scylla.metrics.process import (
-                ChangeResult,
-                ProgressStep,
-                ProgressTracker,
-                calculate_cfp,
-                calculate_pr_revert_rate,
-                calculate_r_prog,
-                calculate_strategic_drift,
-            )
-
-            if progress_tracking and isinstance(progress_tracking, list):
-                steps = []
-                achieved = []
-                for s in progress_tracking:
-                    if isinstance(s, dict):
-                        step = ProgressStep(
-                            step_id=s.get("step_id", ""),
-                            description=s.get("description", ""),
-                            weight=float(s.get("weight", 1.0)),
-                            completed=bool(s.get("completed", False)),
-                            goal_alignment=float(s.get("goal_alignment", 1.0)),
-                        )
-                        steps.append(step)
-                        if step.completed:
-                            achieved.append(step)
-                tracker = ProgressTracker(expected_steps=steps, achieved_steps=achieved)
-                r_prog_val = calculate_r_prog(tracker)
-                strategic_drift_val = calculate_strategic_drift(tracker)
-
-            if changes and isinstance(changes, list):
-                change_results = []
-                for c in changes:
-                    if isinstance(c, dict):
-                        change_results.append(
-                            ChangeResult(
-                                change_id=c.get("change_id", ""),
-                                description=c.get("description", ""),
-                                succeeded=bool(c.get("succeeded", True)),
-                                caused_failure=bool(c.get("caused_failure", False)),
-                                reverted=bool(c.get("reverted", False)),
-                            )
-                        )
-                cfp_val = calculate_cfp(change_results)
-                pr_revert_rate_val = calculate_pr_revert_rate(change_results)
-
-    # Validate and coerce all numeric/boolean fields with type checking
-    return RunData(
-        experiment=experiment,
-        agent_model=agent_model,
-        tier=tier,
-        subtest=subtest,
-        run_number=run_number,
-        score=validate_numeric(result.get("judge_score"), "judge_score", np.nan),
-        passed=validate_bool(result.get("judge_passed"), "judge_passed", False),
-        grade=result.get("judge_grade", "F"),  # String, no validation needed
-        cost_usd=validate_numeric(result.get("cost_usd"), "cost_usd", np.nan),
-        duration_seconds=validate_numeric(
-            result.get("duration_seconds"), "duration_seconds", np.nan
-        ),
-        agent_duration_seconds=validate_numeric(
-            result.get("agent_duration_seconds"), "agent_duration_seconds", np.nan
-        ),
-        judge_duration_seconds=validate_numeric(
-            result.get("judge_duration_seconds"), "judge_duration_seconds", np.nan
-        ),
-        token_stats=token_stats,
-        exit_code=validate_int(result.get("exit_code"), "exit_code", -1),
-        judges=judges,
-        api_calls=api_calls_val,
-        num_turns=num_turns_val,
-        model_usage=model_usage_val,
-        r_prog=r_prog_val,
-        strategic_drift=strategic_drift_val,
-        cfp=cfp_val,
-        pr_revert_rate=pr_revert_rate_val,
-    )
-
-
-def load_experiment(
-    experiment_dir: Path,
-    agent_model: str,
-    experiment_name: str | None = None,
-) -> list[RunData]:
-    """Load all runs from an experiment.
-
-    Args:
-        experiment_dir: Path to experiment directory (contains tier dirs)
-        agent_model: Agent model display name
-        experiment_name: Override for experiment name. If None, uses directory name.
-
-    Returns:
-        List of all run data
-
-    Note:
-        Automatically skips non-tier directories (config, judges.txt, etc.)
-
-    """
-    runs = []
-    experiment_name = experiment_name or experiment_dir.name
-
-    # Iterate through tier directories (T0-T6) — results live under completed/
-    from scylla.e2e.paths import COMPLETED_DIR
-
-    completed_dir = experiment_dir / COMPLETED_DIR
-    scan_base = completed_dir if completed_dir.exists() else experiment_dir
-    for tier_dir in sorted(scan_base.iterdir()):
-        if not tier_dir.is_dir() or not tier_dir.name.startswith("T"):
-            continue
-
-        tier_id = tier_dir.name
-
-        # Iterate through subtest directories
-        for subtest_dir in sorted(tier_dir.iterdir()):
-            if not subtest_dir.is_dir() or not subtest_dir.name.isdigit():
-                continue
-
-            subtest_id = subtest_dir.name
-
-            # Iterate through run directories
-            for run_dir in sorted(subtest_dir.iterdir()):
-                if not run_dir.is_dir() or not run_dir.name.startswith("run_"):
-                    continue
-
-                try:
-                    run = load_run(run_dir, experiment_name, tier_id, subtest_id, agent_model)
-                    runs.append(run)
-                except Exception as e:
-                    logger.warning("Failed to load %s: %s", run_dir, e)
-                    continue
-
-    return runs
-
-
-def load_all_experiments(
-    data_dir: Path,
-    exclude: list[str] | None = None,
-) -> dict[str, list[RunData]]:
-    """Load all experiments from a data directory.
-
-    Args:
-        data_dir: Path to fullruns directory
-        exclude: List of experiment names to exclude (default: [])
-
-    Returns:
-        Dictionary mapping experiment name to list of runs
-
-    Note:
-        To load rubric weights with conflict resolution, call
-        :func:`load_rubric_weights` separately with the desired
-        ``rubric_conflict`` policy.
-
-    """
-    if exclude is None:
-        exclude = []
-
-    experiments = {}
-
-    for exp_dir in sorted(data_dir.iterdir()):
-        if not exp_dir.is_dir():
-            continue
-
-        # Find the timestamped subdirectory (use latest if multiple)
-        timestamped_dirs = sorted([d for d in exp_dir.iterdir() if d.is_dir()])
-        if not timestamped_dirs:
-            continue
-
-        # Use the latest timestamped directory (sorted alphabetically = chronologically)
-        actual_exp_dir = timestamped_dirs[-1]
-        exp_name = exp_dir.name
-
-        if exp_name in exclude:
-            logger.info("Skipping excluded experiment: %s", exp_name)
-            continue
-
-        logger.info("Loading experiment: %s", exp_name)
-
-        # Resolve agent model from experiment configuration
-        try:
-            agent_model = resolve_agent_model(actual_exp_dir)
-        except ValueError as e:
-            logger.warning("%s, skipping experiment", e)
-            continue
-
-        runs = load_experiment(actual_exp_dir, agent_model, experiment_name=exp_name)
-        experiments[exp_name] = runs
-        logger.info("  Loaded %d runs (agent model: %s)", len(runs), agent_model)
-
-    return experiments
-
-
-def load_rubric_weights(  # noqa: C901  # config loading with many format/version branches
-    data_dir: Path,
-    exclude: list[str] | None = None,
-    rubric_conflict: RubricConflict = "error",
-) -> dict[str, float]:
-    """Load and merge category weights from all experiments' rubric.yaml files.
-
-    Scans every experiment directory for rubric.yaml and parses
-    ``categories.*.weight``.  When the same category appears in more than one
-    experiment the ``rubric_conflict`` policy controls resolution:
-
-    * ``'error'`` (default) – raise :class:`RubricConflictError` immediately.
-    * ``'warn'``  – emit a :class:`UserWarning` and keep the *first* value.
-    * ``'first'`` – silently keep the first value encountered.
-    * ``'last'``  – silently overwrite with the most-recently-seen value.
-
-    Float comparison uses a tolerance of ``1e-6`` to avoid spurious conflicts
-    from JSON serialisation round-trips.
-
-    Args:
-        data_dir: Root fullruns directory.
-        exclude: List of experiment names to exclude.
-        rubric_conflict: Policy for handling conflicting rubric weights.
-
-    Returns:
-        Dictionary mapping category names to weights, or ``{}`` if no
-        rubric.yaml was found in any experiment.
-
-    Raises:
-        RubricConflictError: If ``rubric_conflict='error'`` and a conflict is
-            detected.
-
-    """
-    exclude = exclude or []
-
-    # accumulated_weights maps category → (weight, source_experiment_name)
-    accumulated: dict[str, tuple[float, str]] = {}
-    found_any = False
-
-    for exp_dir in sorted(data_dir.iterdir()):
-        if not exp_dir.is_dir() or exp_dir.name in exclude:
-            continue
-
-        exp_name = exp_dir.name
-
-        # Use the latest timestamp directory (sorted alphabetically = chronologically)
-        ts_dirs = sorted(d for d in exp_dir.iterdir() if d.is_dir())
-        if not ts_dirs:
-            continue
-        ts_dir = ts_dirs[-1]
-
-        rubric_path = ts_dir / "rubric.yaml"
-        if not rubric_path.exists():
-            continue
-
-        with rubric_path.open() as f:
-            data = yaml.safe_load(f)
-
-        categories = data.get("categories", {}) if data else {}
-        found_any = True
-
-        for cat_name, cat_data in categories.items():
-            new_weight: float = cat_data.get("weight", 0.0) if cat_data else 0.0
-
-            if cat_name not in accumulated:
-                accumulated[cat_name] = (new_weight, exp_name)
-                continue
-
-            existing_weight, existing_exp = accumulated[cat_name]
-            if abs(existing_weight - new_weight) <= 1e-6:
-                # Identical within tolerance – no conflict.
-                continue
-
-            # Genuine conflict – apply policy.
-            if rubric_conflict == "error":
-                raise RubricConflictError(
-                    category=cat_name,
-                    exp_first=existing_exp,
-                    weight_first=existing_weight,
-                    exp_second=exp_name,
-                    weight_second=new_weight,
-                )
-            elif rubric_conflict == "warn":
-                warnings.warn(
-                    f"Rubric conflict for category '{cat_name}': "
-                    f"experiment '{existing_exp}' defines weight={existing_weight}, "
-                    f"but experiment '{exp_name}' defines weight={new_weight}. "
-                    "Keeping first value.",
-                    UserWarning,
-                    stacklevel=2,
-                )
-                # Keep first – no update to accumulated.
-            elif rubric_conflict == "first":
-                pass  # Keep first – no update to accumulated.
-            elif rubric_conflict == "last":
-                accumulated[cat_name] = (new_weight, exp_name)
-
-    if not found_any:
-        return {}
-
-    return {cat: weight for cat, (weight, _) in accumulated.items()}
+from scylla.analysis.loader_internals.experiment_loader import (
+    load_all_experiments as load_all_experiments,
+)
+from scylla.analysis.loader_internals.experiment_loader import (
+    load_experiment as load_experiment,
+)
+from scylla.analysis.loader_internals.experiment_loader import (
+    load_rubric_weights as load_rubric_weights,
+)
+from scylla.analysis.loader_internals.models import (
+    CriterionScore as CriterionScore,
+)
+from scylla.analysis.loader_internals.models import (
+    ItemScore as ItemScore,
+)
+from scylla.analysis.loader_internals.models import (
+    JudgeEvaluation as JudgeEvaluation,
+)
+from scylla.analysis.loader_internals.models import (
+    ModelUsage as ModelUsage,
+)
+from scylla.analysis.loader_internals.models import (
+    RubricConflict as RubricConflict,
+)
+from scylla.analysis.loader_internals.models import (
+    RubricConflictError as RubricConflictError,
+)
+from scylla.analysis.loader_internals.models import (
+    RunData as RunData,
+)
+from scylla.analysis.loader_internals.run_loader import (
+    load_agent_result as load_agent_result,
+)
+from scylla.analysis.loader_internals.run_loader import (
+    load_judgment as load_judgment,
+)
+from scylla.analysis.loader_internals.run_loader import (
+    load_run as load_run,
+)
+from scylla.analysis.loader_internals.run_loader import (
+    model_id_to_display as model_id_to_display,
+)
+from scylla.analysis.loader_internals.run_loader import (
+    parse_judge_model as parse_judge_model,
+)
+from scylla.analysis.loader_internals.run_loader import (
+    resolve_agent_model as resolve_agent_model,
+)
+from scylla.analysis.loader_internals.validators import (
+    validate_bool as validate_bool,
+)
+from scylla.analysis.loader_internals.validators import (
+    validate_int as validate_int,
+)
+from scylla.analysis.loader_internals.validators import (
+    validate_numeric as validate_numeric,
+)
+
+__all__ = [
+    "CriterionScore",
+    "ItemScore",
+    "JudgeEvaluation",
+    "ModelUsage",
+    "RubricConflict",
+    "RubricConflictError",
+    "RunData",
+    "load_agent_result",
+    "load_all_experiments",
+    "load_experiment",
+    "load_judgment",
+    "load_rubric_weights",
+    "load_run",
+    "model_id_to_display",
+    "parse_judge_model",
+    "resolve_agent_model",
+    "validate_bool",
+    "validate_int",
+    "validate_numeric",
+]
